Added new tcp and file-i/o utility funcions. Refactored code so that all functions are defined within two headers ncph.cpp and tcpUtils.c. Compiling will require new g++ compile directives.
