+++ src/ncphserver.cpp
@@ -17,100 +17,8 @@
     For any questions, you may contact NCP-Hash Group via opening an issue on https://github.com/ncp-hash/public-phash/issues
 */
 
-#include <iostream>
-#include <math.h>
-#include <cstdlib>
-#include <stdio.h>
-#include <random>
-#include<vector>
-#include <gmpxx.h>
-#include<fstream>
-#include<sstream>
-#include"assert.h"
-extern "C" {
-    #include <gmp.h>
-    #include <paillier.h>
-}
-#include <chrono>
-#include <thread>
-#include <string>
-
-int get_rand_index(int const &size) {
-    /* PRE: accepts the size of the array for which random index is needed, must be positive int
-        POST: returns a random index that follows uniform distribution */
-
-    std::random_device rd;                                  // obtain a different random number from hardware
-    std::mt19937 eng(rd());                                 // seed the generator
-    std::uniform_int_distribution<> distr(0, size-1);       // instance of class std::uniform_int_distribution with specific min and max
-    return distr(eng);                                      // obtain a random index for the sample array
-}
-
-float get_beta(float mean, float std) {
-    /* PRE: Accepts the mean and std for normal distribution
-        POST: Returns a single normally distributed variable beta */
-
-    std::random_device rd;                                  // obtain a random number from hardware
-    std::mt19937 gen(rd());                                 // Mersenne twister PRNG, initialized with seed from previous random device instance
-    
-    float sample[1000];
-    for (int i = 0; i < 1000; ++i) {
-        std::normal_distribution<float> d(mean, std);       // instance of class std::normal_distribution with specific mean and stddev
-        sample[i] = d(gen);                                 // get random number with normal distribution using gen as random source
-    }
-    
-    int index = get_rand_index(1000);
-    return sample[index];
-}
-
-int get_rho(int min_rho, int max_rho) {
-    /* PRE: Accepts min and max values for uniform distribution
-        POST: Returns a uniformly distributed variable min <= rho <= max */
-
-    std::random_device rd;                                  // obtain a random number from hardware
-    std::mt19937 gen(rd());                                 // Mersenne twister PRNG, initialized with seed from previous random device instance
-    
-    float sample[1000];
-    for (int i = 0; i < 1000; ++i) {
-        std::uniform_int_distribution<> d(min_rho, max_rho);// instance of class std::normal_distribution with specific mean and stddev
-        sample[i] = d(gen);                                 // get random number with normal distribution using gen as random source
-    }
-    
-    int index = get_rand_index(1000);
-    return sample[index];
-}
-
-void get_rho_beta_arr (int num_rhos, std::vector<float> &betas, int min_rho, int max_rho, float beta_mean, float beta_std) {
-    /* PRE: Accepts number or radii to copute corresponding betas for
-        POST: Returns 256 integers long array of corresponding betas with beta(i) if there exists rho==i and 0 o.w. */
-
-    std::vector<int> rhos;                                  // initialize vector to store rhos
-    int rho = get_rho(min_rho, max_rho);                    // generate first rho - push it to rhos vector and get beta for index=rho in betas vector
-    rhos.push_back(rho);
-    betas[rho] = get_beta(beta_mean, beta_std);
-    
-    for (int i = 1; i < num_rhos; ++i) {                    // repeat the above with check for uniqueness of rhos:
-        while(1) { 
-            rho = get_rho(min_rho, max_rho);                // get random rho
-            if (std::find(rhos.begin(), rhos.end(), rho) != rhos.end()) { continue; } // if rho is not unique, try again
-            rhos.push_back(rho);                            // otherwise, push to rho vector and get beta for index=rho in betas vector
-            betas[rho] = get_beta(beta_mean, beta_std);
-            break;
-        }
-    }
-}
-
-long long int check_sum(std::vector<float> const &betas, std::vector<int> const &rho_sums) {
-    /* PRE: Accepts vector of flots for beta and vector of integers for sums of pixels for all rhos, must be same size
-        POST: Returns a hash value based on scheme 2 with all computations in the clear */
-    
-    assert(betas.size() == rho_sums.size());
-    int limit = betas.size();
-    long long int sum = 0;
-    for (int i = 0; i < limit; ++i) {
-        sum += ((int)abs(betas[i])) * rho_sums[i];
-    }
-    return sum;
-}
+#include "ncph.hpp"
+#include "tcpUtils.c"
 
 int main() {
     /* PARAMETERS & CONSTRAINTS:
