+++ src/tcpUtils.c
@@ -10,17 +10,17 @@
 // server_recieve_enc_hash
 
 //******************************************************************************************
-
+#include <assert.h>
 
 #include <stdio.h> 
 #include <unistd.h>
-
 #include <netdb.h> 
 #include <netinet/in.h> 
 #include <stdlib.h> 
 #include <string.h> 
 #include <sys/socket.h> 
 #include <sys/types.h> 
+#include <arpa/inet.h>
 
 // #define MAX 513 //256, 2-byte betas with one /n character at end 
 #define MAX 65537 // 65,536 bytes plus one terminating byte
@@ -29,9 +29,18 @@
 #define TRUE 1
 #define FALSE 0
 
-#define SUCCESSS 0
+#define SUCCESS 0
 #define ABORT 1
 
+#define BETA_SIZE 256 //each beta is 256 bytes
+#define NUM_BETAS 256 //there are 256 betas in total
+#define PAILLIER_KEY_SIZE 256 //this is 256 bytes
+
+int min(int first, int second){
+
+	if(first<second) return first;
+	else if(first>=second) return second;
+}
 
 int server_init(void){
 
@@ -53,7 +62,7 @@ int server_init(void){
 	serverAddress.sin_addr.s_addr = htonl(INADDR_ANY); 
 	serverAddress.sin_port = htons(PORT); 
 
-	if ((bind(listening_socket_fd, (struct sockaddr*)&serverAddress, sizeof(serverAddress))) != 0) { 
+	if ((bind(listening_socket_fd, (struct sockaddr*)&serverAddress, sizeof(serverAddress))) != SUCCESS) { 
 		printf("socket bind failure\n"); 
 		exit(0); 
 	} 
@@ -61,7 +70,7 @@ int server_init(void){
 		printf("socket bind success\n"); 
 
 
-	if ((listen(listening_socket_fd, 5)) != 0) { 
+	if ((listen(listening_socket_fd, 5)) != SUCCESS) { 
 		printf("listen failure\n"); 
 		exit(0); 
 	} 
@@ -77,6 +86,7 @@ int server_connect_to_client(int listening_socket_fd){
 
 	struct sockaddr_in clientAddress; 
 	unsigned int len; 
+	int client_socket_fd;
 
 	len = sizeof(clientAddress); 
 	client_socket_fd = accept(listening_socket_fd, (struct sockaddr*)&clientAddress, &len);
@@ -124,7 +134,6 @@ int client_connect_to_server(void){
 	return socketFD;
 }
 
-
 //sends message, blocks for response. 
 // will loop indefinitely without no recipient response
 //re implement with forks
@@ -132,7 +141,9 @@ int client_connect_to_server(void){
 int send_char_string(int recipient_socket_fd, char* message, int message_len){
 
 	char recipient_response[message_len];
-	char expected_response[message_len] = "message received.\n";
+	char* expected_response = (char*)malloc((message_len+1) * sizeof(char));
+	strcpy(expected_response, "message received.\n"); //TODO: may need to use strncopy 
+	//char expected_response[message_len] = "message received.\n";
 
 	while(TRUE){
 
@@ -140,10 +151,10 @@ int send_char_string(int recipient_socket_fd, char* message, int message_len){
 		printf("message sent:\n%s\n", message);
 
 		bzero(recipient_response, message_len); 
-		read(socketFD, recipient_response, message_len);
+		read(recipient_socket_fd, recipient_response, message_len);
 
 		if ((strcmp(recipient_response, expected_response)) == 0) { 
-			printf("message received.\n"); 
+			printf("message was confirmed as received.\n"); 
 			return 0; 
 		} 
 	} 
@@ -151,171 +162,250 @@ int send_char_string(int recipient_socket_fd, char* message, int message_len){
 }
 
 
-void receive_char_string(int sender_socket_fd, char* message){
+//recieves a character string and then returns a confirmation to sender
+char* receive_char_string(int sender_socket_fd, int message_len){
 
-	char clientEncHash[MAX]; 
-	char confirmation[MAX] = "message received.\n";
+	char* received_message = (char*)malloc((message_len+1) * sizeof(char));
+
+	char* confirmation = (char*)malloc((message_len+1) * sizeof(char));
+	strcpy(confirmation, "message received.\n"); //TODO: may need to use strncopy 
+
 	char endOfTransmission;
 
 	while (TRUE) { 
 
-		bzero(clientEncHash, MAX); 
+		bzero(received_message, message_len); 
 
-		// get message from client; copy it to clientEncHash buffer 
-		read(sender_socket_fd, clientEncHash, MAX); 
-		
+		// get message from client; copy it to received_message buffer 
+		read(sender_socket_fd, received_message, message_len); 
+		printf("message received:\n%s\n", received_message);
+
 		// and send that confirmation message to client 
-		write(sender_socket_fd, confirmation, MAX); //may want to replace with rec or send avoid problems in large filetypes
+		write(sender_socket_fd, confirmation, message_len); //may want to replace with rec or send avoid problems in large filetypes
 
 		// if client string contains "\n" then server exit and chat ended.
-		//endOfTransmission = clientEncHash[strlen(clientEncHash)-1]; 
-		endOfTransmission = clientEncHash[MAX-1]; 
+		//endOfTransmission = received_message[strlen(received_message)-1]; 
+		endOfTransmission = received_message[message_len-1]; 
 		if (strcmp("\n", &endOfTransmission) == 0) { 
 			printf("server quitting\n"); 
 			break; 
 		}
 	} 
-	/************************* 
-		need to return client
-		input here
-	**************************/
 
+	free(confirmation);
+	return received_message;
 }
 
-//returns 0 on success (positive confirmation)
-int server_send_paillier_pubkey(int client_socket_fd, char* paillier_pubkey, int paillier_pubkey_len){
+// ========================================================================================
+// KeyFile Read and Write Utilities
+// ========================================================================================
 
-	ret_val = send_char_string(client_socket_fd, paillier_pubkey, paillier_pubkey_lenl);
-	return ret_val;
-} 
+// These are written in c because there will be a parent process initialize the keyfiles and
+// to start up the hash recieving server. This parent process will be written in c and 
+// will use posix fork() calls.
 
-int send_bytes_chunk(int recipient_socket_fd, void *chunk_buffer, int chunk_len){
+// When the client and server are run on the same computer, the parent will first initialize
+// the key files which both child processes (the hash recieving server and the client) will
+// read from, but not write to. This solves a race condition issue experienced in a previous
+// implementation, while also better simulating the separate nature of the client and server
+// given that they'd be separate processes that will communicate exclusively through tcp.
 
-    unsigned char *pbuf = (unsigned char *) chunk_buffer;
+// when client and server are on separate devices, both will still have access to the keyfiles.
+// the client will have the keyfiles as part it's installation files and the server will hve a
+// copy locally (since it generated the public keys in the first place.)
 
-    while (chunk_len > 0){
+//what is the input type of the key itself? Is it void?
+void write_key_file(char* file_name, void* key){
 
-        int num = send(recipient_socket_fd, pbuf, chunk_len, 0);
-        if (num == SOCKET_ERROR){
+	int num_bytes;
+    if (!strncmp(file_name,"betas",5)) num_bytes = (BETA_SIZE * NUM_BETAS);
+    if (!strncmp(file_name,"paillier",8)) num_bytes = PAILLIER_KEY_SIZE;
+    if (!strncmp(file_name,"betas_test",10)) num_bytes = (BETA_SIZE * NUM_BETAS);
 
-            if (WSAGetLastError() == WSAEWOULDBLOCK){
 
-                // optional: use select() to check for timeout to fail the send
-                continue;
-            }
-            return ABORT;
-        }
+	//intialize filename
+	char file_path[255];
+	snprintf(file_path,255,"./%s.key",file_name);
+	
+	//open file in append mode if a single beta
+	FILE *fptr;
+	if(!strncmp(file_name,"betas",5)) fptr = fopen(file_path,"ab");
+	else fptr = fopen(file_path,"wb");
 
-        pbuf += num;
-        chunk_len -= num;
-    }
+	//write null terminator when it's a paillier key (may need null terminator for both)
+	fwrite(key,1,num_bytes,fptr);
+	if(!strncmp(file_name,"paillier",8)) fwrite("\0",1,1,fptr); //write the string null terminator to file
 
-    return SUCCESS;
+	if(fptr == NULL)
+	{
+	  printf("Error!");   
+	  exit(1);             
+	}
+
+	fclose(fptr);
 }
 
 
-int send_bytes_size(int recipient_socket_fd, long bytes_size){
+//to cast the betas returned from this function in the c++ code, do the following:
+// std::vector<paillier_ciphertext_t> vector_name(ptr_to_return_key, ptr_to_return_key + number_of_elements);
+// notice that the offset in the last argument is for the number of elements NOT the number of bytes.
 
-    bytes_size = htonl(bytes_size);
-    return send_bytes_chunk(recipient_socket_fd, &bytes_size, sizeof(bytes_size));
+//to cast the paillier public key returned from this function in c++ code, do this:
+// paillier_ciphertext_t paillier_key = *ptr_to_return_key;
+
+void* read_key_file(char* file_name){
+
+	int num_bytes;
+    if (!strncmp(file_name,"betas",5)) num_bytes = (BETA_SIZE * NUM_BETAS);
+    if (!strncmp(file_name,"paillier",8)) num_bytes = PAILLIER_KEY_SIZE + 1;// +1 byte for the null terminator
+
+	//initialize filename
+	char file_path[255];
+	snprintf(file_path,255,"./%s.key",file_name);
+
+	//alocate memory for key to be returned
+    FILE *key_file = fopen(file_path, "rb");  
+    unsigned char *return_key = malloc(num_bytes); //allocates that many bytes
+    
+    //read key from file and close
+    fread(return_key, 1, num_bytes, key_file);
+    fclose(key_file); 
+
+    //return the paillier public key OR the pailler ciphertext (betas) depending on file_name
+	if (!strncmp(file_name,"betas",5)) return paillier_ciphertext_from_bytes(return_key, num_bytes);
+	if (!strncmp(file_name,"paillier",8)) return paillier_pubkey_from_hex(return_key);
+	return 1;
 }
 
+// int send_bytes_chunk(int recipient_socket_fd, void *chunk_buffer, int chunk_len){
 
-int send_bytes_all(int recipient_socket_fd, void* all_bytes, int all_bytes_len){
+//     unsigned char *pbuf = (unsigned char *) chunk_buffer; //todo: check on this later
 
-    unsigned char *bytes_buffer = (unsigned char *) all_bytes;
+//     while (chunk_len > 0){
+
+//         int num = send(recipient_socket_fd, pbuf, chunk_len, 0);
+//         if (num == -1) return ABORT;
+
+//         pbuf += num;
+//         chunk_len -= num;
+//     }
+
+//     return SUCCESS;
+// }
 
 
-    if (send_bytes_len(recipient_socket_fd, all_bytes_len) == ABORT)
-        return ABORT;
+// int send_bytes_len(int recipient_socket_fd, long bytes_len){
 
-    if (all_bytes_len > 0)
-    {
-        char chunk_buffer[1024];
-        do
-        {
-            unsigned int num_bytes_read = min(all_bytes_len, sizeof(chunk_buffer));
+//     bytes_len = htonl(bytes_len);
+//     return send_bytes_chunk(recipient_socket_fd, &bytes_len, sizeof(bytes_len));
+// }
 
-            if (num_bytes_read < 1)
-                return ABORT;
+
+// int send_bytes_all(int recipient_socket_fd, void* all_bytes, int all_bytes_len){
+
+//     unsigned char *bytes_buffer = (unsigned char *) all_bytes;
+
+
+//     if (send_bytes_len(recipient_socket_fd, all_bytes_len) == ABORT)
+//         return ABORT;
+
+//     if (all_bytes_len > 0)
+//     {
+//         char chunk_buffer[1024];
+//         do
+//         {
+
+//             unsigned int num_bytes_read = min(all_bytes_len, sizeof(chunk_buffer));
+
+//             if (num_bytes_read < 1){
+//                 return ABORT;
+//                 printf("ABORT SEND\n");
+//             }
+
             
-            strncpy(chunk_buffer, bytes_buffer, num_bytes_read);
+//             strncpy(chunk_buffer, bytes_buffer, num_bytes_read);
 
-            if (send_bytes_chunk(recipient_socket_fd, chunk_buffer, num_bytes_read) == ABORT)
-                return ABORT;
+//             if (send_bytes_chunk(recipient_socket_fd, chunk_buffer, num_bytes_read) == ABORT)
+//                 return ABORT;
 
-            all_bytes_len -= num_bytes_read;
-            all_bytes += num_bytes_read;
-        }
-        while (all_bytes_len > 0);
+//             all_bytes_len -= num_bytes_read;
+//             all_bytes += num_bytes_read;
+//         }
+//         while (all_bytes_len > 0);
 
-    }
-    return SUCCESS;
-}
+//     }
 
+//     for(int i = 0; i < 256*256; i++){
+//         printf("%02x", ((char *)all_bytes)[i]);
+//     }
+//     putchar( '\n' );
 
-
-int receive_bytes_chunk(int sender_socket_fd, void *chunk_buffer, int chunk_buffer_len){
-
-    unsigned char *pchunk_buffer = (unsigned char *) chunk_buffer;
-
-    while (chunk_buffer_len > 0){
-
-        int num_bytes_received = recv(sender_socket_fd, pchunk_buffer, chunk_buffer_len, 0);
-        if (num_bytes_received == SOCKET_ERROR){
-
-            if (WSAGetLastError() == WSAEWOULDBLOCK){
-
-                // optional: use select() to check for timeout to fail the read
-                continue;
-            }
-            return ABORT;
-        }
-        else if (num_bytes_received == 0)
-            return ABORT;
-
-        pchunk_buffer += num_bytes_received;
-        chunk_buffer_len -= num_bytes_received;
-    }
-
-    return SUCCESS;
-}
+//     return SUCCESS;
+// }
 
 
-int receive_bytes_len(int sender_socket_fd, long *bytes_len){
 
-    if (receive_bytes_chunk(sender_socket_fd, bytes_len, sizeof(bytes_len)) == ABORT)
-        return ABORT;
+// int receive_bytes_chunk(int sender_socket_fd, void *chunk_buffer, int chunk_buffer_len){
 
-    *bytes_len = ntohl(*bytes_len);
-    return SUCCESS;
-}
+//     unsigned char *pchunk_buffer = (unsigned char *) chunk_buffer;
 
-int receive_bytes_all(int sender_socket_fd, void* all_bytes){
+//     while (chunk_buffer_len > 0){
 
-    unsigned char *bytes_buffer = (unsigned char *) all_bytes;
+//         int num_bytes_received = recv(sender_socket_fd, pchunk_buffer, chunk_buffer_len, 0);
 
-    long all_bytes_len;
-    long offset = 0;
+//         if (num_bytes_received == -1 || num_bytes_received == 0) return ABORT; 
 
-    if (receive_bytes_len(sender_socket_fd, &all_bytes_len) == ABORT)
-        return ABORT;
+//         pchunk_buffer += num_bytes_received;
+//         chunk_buffer_len -= num_bytes_received;
+//     }
 
-    if (all_bytes_len > 0){
+//     return SUCCESS;
+// }
 
-        char chunk_buffer[1024];
 
-        do{
-            int num_bytes_received = min(all_bytes_len, sizeof(chunk_buffer));
+// int receive_bytes_len(int sender_socket_fd, long *bytes_len){
 
-            if (receive_bytes_chunk(sender_socket_fd, chunk_buffer, num_bytes_received) == ABORT)
-                return ABORT;
+//     if (receive_bytes_chunk(sender_socket_fd, bytes_len, sizeof(bytes_len)) == ABORT)
+//         return ABORT;
 
-            memcopy(all_bytes + offset, chunk_buffer, num_bytes_received)
-            offset += num_bytes_received
-            all_bytes_len -= num_bytes_received;
+//     *bytes_len = ntohl(*bytes_len);
+//     return SUCCESS;
+// }
 
-        } while (all_bytes_len > 0);
-    }
-    return SUCCESS;
-}
+// int receive_bytes_all(int sender_socket_fd, void* all_bytes){
+
+// 	printf("receive was called\n");
+
+//     unsigned char *bytes_buffer = (unsigned char *) all_bytes;
+
+//     long all_bytes_len;
+//     long offset = 0;
+
+//     if (receive_bytes_len(sender_socket_fd, &all_bytes_len) == ABORT)
+//         return ABORT;
+
+//     if (all_bytes_len > 0){
+
+//         char chunk_buffer[1024];
+//         do{
+//             int num_bytes_received = min(all_bytes_len, sizeof(chunk_buffer));
+
+//             if (receive_bytes_chunk(sender_socket_fd, chunk_buffer, num_bytes_received) == ABORT)
+//                 return ABORT;
+
+//             memcpy(all_bytes + offset, chunk_buffer, num_bytes_received);
+//             offset += num_bytes_received;
+//             all_bytes_len -= num_bytes_received;
+
+//         } while (all_bytes_len > 0);
+//     }
+
+
+//     //print out the 
+//     for(int i = 0; i < 256*256; i++){
+//         printf("%02x", ((char *) all_bytes)[i]);
+//     }
+//     putchar( '\n' );
+
+//     return SUCCESS;
+// }
